+++ CHANGELOG.rst
@@ -6,9 +6,15 @@ This document describes changes between each past release.
 3.4.0 (unreleased)
 ------------------
 
+**Bug fixes**
+
+- Prevent events to be sent if status is not changed (#268)
+
 **Internal Changes**
 
+- Rewrite e2e to use ``.get_records_timestamp()`` (#258)
 - Enable Wheel distribution (fixes #271)
+
 
 3.3.3 (2018-10-10)
 ------------------